requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
ffmpeg-downloader==0.3.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
jiter==0.8.2
numpy==2.0.2
openai==1.59.4
packaging==24.2
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.1
sentencepiece==0.2.0
sniffio==1.3.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2024.3.10
